--- client/src/view/components/Carousel/config/default.js
@@ -1,17 +1,0 @@
-const slide = {
-  className: 'swiper__slide',
-};
-
-const carousel = {
-  loop: true,
-  effect: 'cube',
-  pagination: {
-    clickable: true,
-  },
-  autoplay: {
-    delay: 5000,
-  },
-  navigation: true,
-};
-
-export default { slide, carousel };
