+++ client/src/view/components/Carousel/config/default.jsx
@@ -1,6 +1,22 @@
-import { createUseStyles } from 'react-jss';
+import { makeStyles } from '@material-ui/core/styles';
 
-const styles = {
+const slide = {
+  className: 'swiper__slide',
+};
+
+const carousel = {
+  loop: true,
+  effect: 'cube',
+  pagination: {
+    clickable: true,
+  },
+  autoplay: {
+    delay: 5000,
+  },
+  navigation: true,
+};
+
+const styles = makeStyles({
   caroulesWrapper: {
     width: '100%',
     display: 'flex',
@@ -54,6 +70,6 @@ const styles = {
       mixBlendMode: 'exclusion',
     },
   },
-};
+});
 
-export default createUseStyles(styles);
+export default { slide, carousel, styles };
