+++ client/src/view/components/Carousel/index.jsx
@@ -1,5 +1,5 @@
 import Carousel from './Carousel';
 
 export default Carousel;
-export { default as defaultCarousel } from './config/default.js';
-export { default as popularProductCard } from './config/custom.js';
+export { default as defaultCarousel } from './config/default';
+export { default as popularProductCard } from './config/custom';
